Make sure our external api is separated from the internal in Swagger/Open API.
