+++ src/main/kotlin/no/nav/klage/oppgave/api/controller/external/ExternalApiController.kt
@@ -1,4 +1,4 @@
-package no.nav.klage.oppgave.api.controller
+package no.nav.klage.oppgave.api.controller.external
 
 import io.swagger.v3.oas.annotations.Operation
 import io.swagger.v3.oas.annotations.Parameter
